+++ cmd2/argparse_completer.py
@@ -654,12 +654,19 @@ class AutoCompleter(object):
             else:
                 prefix = ''
 
+        if action.help is None:
+            help_text = ''
+        else:
+            help_text = action.help
+
+        # is there anything to print for this parameter?
+        if not prefix and not help_text:
+            return
+
         prefix = '  {0: <{width}}    '.format(prefix, width=20)
         pref_len = len(prefix)
-        if action.help is not None:
-            help_lines = action.help.splitlines()
-        else:
-            help_lines = ['']
+        help_lines = help_text.splitlines()
+
         if len(help_lines) == 1:
             print('\nHint:\n{}{}\n'.format(prefix, help_lines[0]))
         else:
@@ -676,12 +683,12 @@ class AutoCompleter(object):
         """
         Performs tab completion against a list
 
-        :param text: str - the string prefix we are attempting to match (all returned matches must begin with it)
-        :param line: str - the current input line with leading whitespace removed
-        :param begidx: int - the beginning index of the prefix text
-        :param endidx: int - the ending index of the prefix text
-        :param match_against: Collection - the list being matched against
-        :return: List[str] - a list of possible tab completions
+        :param text: the string prefix we are attempting to match (all returned matches must begin with it)
+        :param line: the current input line with leading whitespace removed
+        :param begidx: the beginning index of the prefix text
+        :param endidx: the ending index of the prefix text
+        :param match_against: the list being matched against
+        :return: a list of possible tab completions
         """
         return [cur_match for cur_match in match_against if cur_match.startswith(text)]